requirements.txt
streamlit==1.33.0
streamlit-authenticator==0.2.3
pandas==1.5.2
numpy>= 1.21.0, <1.25.0
fastapi==0.110.1
matplotlib==3.8
requests==2.31.0
python-dotenv==1.0.0